requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
psycopg2
python-jose[cryptography]
bcrypt
python-dotenv
passlib